requirements.txt
# Video extraction
yt-dlp>=2024.1.1
youtube-transcript-api>=0.6.1
bilibili-api>=16.0.0

# Speech recognition
openai-whisper>=20231117
faster-whisper>=1.0.0

# Knowledge graph
neo4j>=5.14.0
py2neo>=2021.2.4
networkx>=3.2.1

# Vector search
chromadb>=0.4.18
sentence-transformers>=2.3.1

# NLP
spacy>=3.7.2
transformers>=4.35.0

# LLM
openai>=1.6.0
zhipuai>=2.0.0

# Document generation
markdown2>=2.4.12
python-docx>=1.1.0

# MCP (optional)
mcp>=0.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1
